Startup now reads notes from notes.txt (revised)
Firestore saves timestamps as their own Timestamp objects. This conflicts with my usage of `DateTime`. So I had to convert in the
constructors.

But it turns out `DateTime`s are not hashable, so I had to convert
back and forth to and from strings in the JSON methods to store them